+++ R/factor_analysis.R
@@ -42,8 +42,43 @@ factor_analysis.data.frame <- function(x,
 }
 
 
+#' @rdname principal_components
 #' @export
-factor_analysis.matrix <- factor_analysis.data.frame
+factor_analysis.matrix <- function(x,
+                                   n = "auto",
+                                   rotation = "oblimin",
+                                   n_obs = NULL,
+                                   n_matrix = NULL,
+                                   sort = FALSE,
+                                   threshold = NULL,
+                                   standardize = FALSE,
+                                   ...) {
+  # check if we have a square matrix. in this case, we assume that
+  # the user wants to do a factor analysis on the correlation matrix
+  if ((dim(x)[1] == dim(x)[2]) && is.null(n_obs)) {
+    insight::format_error(
+      "You provided a square matrix, which is assumed to be a correlation matrix. Please specify the number of observations with `n_obs`. If your matrix is not a correlation matrix, please provide a data frame instead."
+    )
+  }
+
+  # the default n.obs argument in `psych::fa()` is `NA`, so we change
+  # our default `NULL` to `NA` to avoid errors
+  if (is.null(n_obs)) {
+    n_obs <- NA
+  }
+
+  factor_analysis.data.frame(
+    as.data.frame(x),
+    n = n,
+    rotation = rotation,
+    sort = sort,
+    threshold = threshold,
+    standardize = standardize,
+    n.obs = n_obs,
+    np.obs = n_matrix,
+    ...
+  )
+}
 
 
 .is_oblique_rotation <- function(rotation) {