+++ tests/testthat/test-factor_analysis.R
@@ -50,6 +50,36 @@ test_that("factor_analysis", {
   # check factor scores
   fc <- factor_scores(out)
   expect_identical(dim(fc), c(32L, 2L))
+
+  # works with correlation matrix
+  skip_if_not_installed("correlation")
+  skip_on_cran() # takes too long on CRAN
+
+  raq_poly <- correlation::correlation(raq_items, method = "polychoric")
+  raq_poly_mtx <- as.matrix(raq_poly) # store correlation matrix
+
+  # needs n_obs
+  expect_error(
+    factor_analysis(raq_poly_mtx, n = 4),
+    regex = "You provided a square matrix"
+  )
+
+  out1 <- factor_analysis(raq_poly_mtx, n = 4, n_obs = 2571)
+  expect_identical(dim(out1), c(23L, 7L))
+  expect_named(
+    out1,
+    c("Variable", "MR1", "MR2", "MR4", "MR3", "Complexity", "Uniqueness")
+  )
+
+  out2 <- factor_analysis(as.matrix(raq_items), n = 4)
+  expect_identical(dim(out2), c(23L, 7L))
+  expect_named(
+    out2,
+    c("Variable", "MR1", "MR2", "MR4", "MR3", "Complexity", "Uniqueness")
+  )
+
+  # roughly equal results
+  expect_equal(out1$MR1, out2$MR1, tolerance = 1e-1)
 })